refactor: simplify import statements for TodoRepository and enhance Todo entity documentation
