+++ dddpy/usecase/todo/complete_todo_usecase.py
@@ -7,7 +7,7 @@ from dddpy.domain.todo.exceptions import (
     TodoNotFoundError,
     TodoNotStartedError,
 )
-from dddpy.domain.todo.repositories.todo_repository import TodoRepository
+from dddpy.domain.todo.repositories import TodoRepository
 from dddpy.domain.todo.value_objects import TodoId, TodoStatus
