+++ dddpy/usecase/todo/delete_todo_usecase.py
@@ -3,7 +3,7 @@
 from abc import ABC, abstractmethod
 
 from dddpy.domain.todo.exceptions import TodoNotFoundError
-from dddpy.domain.todo.repositories.todo_repository import TodoRepository
+from dddpy.domain.todo.repositories import TodoRepository
 from dddpy.domain.todo.value_objects import TodoId
